backend/requirements.txt
fastapi
uvicorn
cryptography
pillow
numpy
PyPDF2
python-multipart
